Refactor graph coordinates to use interfaces
