+++ src/main/java/io/patriot_framework/generator/eventSimulator/simulationPackages/graphFire/RoomTempDataFeed.java
@@ -17,7 +17,7 @@
 package io.patriot_framework.generator.eventSimulator.simulationPackages.graphFire;
 
 import io.patriot_framework.generator.Data;
-import io.patriot_framework.generator.eventSimulator.coordinates.UndirectedGraphCoordinate;
+import io.patriot_framework.generator.eventSimulator.coordinates.graph.UndirectedGraphCoordinate;
 import io.patriot_framework.generator.dataFeed.DataFeed;
 import io.patriot_framework.generator.eventSimulator.eventGenerator.eventBus.EventBusClientBase;
